fix: make number to float to work with lox numbers

+++ src/Expressions.zig
@@ -2,7 +2,7 @@ const Scanner = @import("Scanner.zig");
 const std = @import("std");
 
 pub const ExprId = u32;
-pub const LiteralValue = union(enum) { None, String: []const u8, Number: u32, Bool: bool };
+pub const LiteralValue = union(enum) { None, String: []const u8, Number: f32, Bool: bool };
 
 pub const BinaryExpr = struct { left: ExprId, operator: Scanner.TokenType, right: ExprId };
 pub const UnaryExpr = struct { operator: Scanner.TokenType, right: ExprId };
